+++ Source/EventFlow/Aggregates/DomainEvent.cs
@@ -34,7 +34,7 @@ namespace EventFlow.Aggregates
         public Guid BatchId { get; private set; }
         public TAggregateEvent AggregateEvent { get; private set; }
         public long GlobalSequenceNumber { get; private set; }
-        public string AggregateId { get; private set; }
+        public IIdentity AggregateIdentity { get; private set; }
         public IMetadata Metadata { get; private set; }
         public DateTimeOffset Timestamp { get; private set; }
 
@@ -43,7 +43,7 @@ namespace EventFlow.Aggregates
             IMetadata metadata,
             DateTimeOffset timestamp,
             long globalSequenceNumber,
-            string aggregateId,
+            IIdentity aggregateIdentity,
             int aggregateSequenceNumber,
             Guid batchId)
         {
@@ -51,7 +51,7 @@ namespace EventFlow.Aggregates
             Metadata = metadata;
             Timestamp = timestamp;
             GlobalSequenceNumber = globalSequenceNumber;
-            AggregateId = aggregateId;
+            AggregateIdentity = aggregateIdentity;
             AggregateSequenceNumber = aggregateSequenceNumber;
             BatchId = batchId;
         }
@@ -68,7 +68,7 @@ namespace EventFlow.Aggregates
                 AggregateType.Name,
                 AggregateSequenceNumber,
                 EventType.Name,
-                AggregateId);
+                AggregateIdentity);
         }
     }
 }